add the badge and required on form
